requirements.txt
portia-sdk-python[all] 